Trying to fix dependency cycles...
