+++ src/transactions/instances/index.ts
@@ -7,3 +7,4 @@ export * from "./scriptTransactionArguments";
 export * from "./transactionPayload";
 export * from "./moduleId";
 export * from "./identifier";
+export * from "./transactionArgument";
