+++ src/transactions/instances/signedTransaction.ts
@@ -3,7 +3,8 @@
 
 /* eslint-disable @typescript-eslint/naming-convention */
 
-import { Serializer, Deserializer, Serializable } from "../../bcs";
+import { Deserializer } from "../../bcs/deserializer";
+import { Serializable, Serializer } from "../../bcs/serializer";
 import { TransactionAuthenticator } from "../authenticator/transaction";
 import { RawTransaction } from "./rawTransaction";
